Ajout d'une barre de sélection de l'année scolaire et de l'académie
