+++ app/views/academic/shared/_academy_banner.html.haml
@@ -1,0 +1,9 @@
+.etab-banner
+  .fr-container
+    .fr-grid-row.fr-text--sm
+      .fr-col
+        = "#{selected_school_year} - Académie #{selected_academy}"
+      %div
+        = dsfr_link_to "Changer d'année scolaire", select_school_years_path, class: "fr-text--sm"
+        - if academies.many?
+          = dsfr_link_to "Changer d'académie", select_academy_academic_users_path(current_user), class: "fr-text--sm fr-ml-1w"
