+++ app/views/academic/users/select_academy.html.haml
@@ -7,6 +7,6 @@
       = form_with url: selected_academy_academic_users_path, builder: DsfrFormBuilder do |form|
         .fr-input-group
           = form.label :academy, "Académie", class: 'fr-label'
-          = form.select :academy, @academies, {}, { class: 'fr-select' }
+          = form.select :academy, academies, {}, { class: 'fr-select' }
 
         = form.submit "Continuez avec cette académie", class: 'fr-btn'
